Added new object types: FlameJumper and OrbitingHazard. Need to work out some bugs, including collision and movement issues. Made a big test level. Renamed 'Game' class to 'World' and made it static.

+++ src/unnamed_platformer/app/ImageHelper.java
@@ -97,7 +97,7 @@ public class ImageHelper {
 			return new ImageIcon(img);
 		}
 
-		float ratio = img.getWidth() / img.getHeight();
+		float ratio = (float)img.getWidth() / img.getHeight();
 
 		return new ImageIcon(img.getScaledInstance(size, (int) (size / ratio), java.awt.Image.SCALE_SMOOTH));
 	}
